Avoid killing pydev threads on an exception
Killing run-perf when running from inside eclipse's pydev can result in
X server crash. See:

    https://bugzilla.redhat.com/show_bug.cgi?id=1873489

for details. To avoid that let's simply detect the "pydev.Reader" string
and avoid killing threads when detected.

+++ runperf/__init__.py
@@ -273,9 +273,15 @@ def main():
                 hosts.cleanup()
         # TODO: Treat hanging background threads
         if len(threading.enumerate()) > 1:
-            log.warning("Background threads present, killing: %s",
-                        threading.enumerate())
-            os.kill(0, 15)
+            threads = threading.enumerate()
+            if any("pydevd.Reader" in _ for _ in threads):
+                logging.warning("Background threads %s present but 'pydev' "
+                                "thread detected, not killing anything"
+                                % threads)
+            else:
+                log.warning("Background threads present, killing: %s",
+                            threading.enumerate())
+                os.kill(0, 15)
         raise
